adjusted doc2vec to work with trim_rule
