+++ gensim/utils.py
@@ -11,6 +11,7 @@ This module contains various general utility functions.
 from __future__ import with_statement
 
 import logging
+
 logger = logging.getLogger(__name__)
 
 try:
@@ -1096,13 +1097,32 @@ def prune_vocab(vocab, min_reduce, trim_rule=None):
     Modifies `vocab` in place, returns the sum of all counts that were pruned.
 
     """
-    trim_rule = trim_rule or (lambda word, count, min_count: count >= min_reduce)
     result = 0
     old_len = len(vocab)
     for w in list(vocab):  # make a copy of dict's keys
-        if not trim_rule(w, vocab[w], min_reduce):  # vocab[w] <= min_reduce:
+        if not _keep_vocab_item(w, vocab[w], min_reduce, trim_rule):  # vocab[w] <= min_reduce:
             result += vocab[w]
             del vocab[w]
     logger.info("pruned out %i tokens with count <=%i (before %i, after %i)",
                 old_len - len(vocab), min_reduce, old_len, len(vocab))
     return result
+
+
+RULE_DEFAULT = 0
+RULE_DISCARD = 1
+RULE_KEEP = 2
+
+
+def _keep_vocab_item(word, count, min_count, trim_rule=None):
+    default_res = count >= min_count
+
+    if trim_rule is None:
+        return default_res
+    else:
+        rule_res = trim_rule(word, count, min_count)
+        if rule_res == RULE_KEEP:
+            return True
+        elif rule_res == RULE_DISCARD:
+            return False
+        else:
+            return default_res
